Deduplicate migration and component version gating
The four migrations now derive from a shared EnterspeedMigrationBase that
holds the AsyncMigrationBase/MigrationBase split once, and
EnterspeedJobsComponent is split into version-suffixed partial classes
(IAsyncComponent on Umbraco 17+, IComponent before) with the migration plan
defined a single time.
