+++ src/Enterspeed.Source.UmbracoCms.Base/Components/EnterspeedJobsComponent.cs
@@ -1,14 +1,7 @@
-#if NET10_0_OR_GREATER
-using System.Threading;
-using System.Threading.Tasks;
-#endif
 using Enterspeed.Source.UmbracoCms.Base.Data.Migration;
-using Umbraco.Cms.Core;
-using Umbraco.Cms.Core.Composing;
 using Umbraco.Cms.Core.Migrations;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Infrastructure.Migrations;
-using Umbraco.Cms.Infrastructure.Migrations.Upgrade;
 #if NET5_0
 using Umbraco.Cms.Core.Scoping;
 #else
@@ -17,11 +10,12 @@ using Umbraco.Cms.Infrastructure.Scoping;
 
 namespace Enterspeed.Source.UmbracoCms.Base.Components
 {
-#if NET10_0_OR_GREATER
-    public class EnterspeedJobsComponent : IAsyncComponent
-#else
-    public class EnterspeedJobsComponent : IComponent
-#endif
+    /// <summary>
+    /// Runs the EnterspeedJobs database migration plan at startup. The component contract
+    /// differs between Umbraco generations (IComponent was removed in Umbraco 18), so the
+    /// interface implementations live in the version-suffixed partial files.
+    /// </summary>
+    public partial class EnterspeedJobsComponent
     {
         private readonly IScopeProvider _scopeProvider;
         private readonly IKeyValueService _keyValueService;
@@ -40,12 +34,8 @@ namespace Enterspeed.Source.UmbracoCms.Base.Components
             _runtimeState = runtimeState;
         }
 
-#if NET10_0_OR_GREATER
-        public async Task InitializeAsync(bool isRestarting, CancellationToken cancellationToken)
+        private static MigrationPlan BuildMigrationPlan()
         {
-            if (_runtimeState.Level < RuntimeLevel.Run)
-                return;
-
             var migrationPlan = new MigrationPlan("EnterspeedJobs");
             migrationPlan.From(string.Empty)
                 .To<EnterspeedJobsTableMigration>("enterspeedjobs-db")
@@ -53,34 +43,7 @@ namespace Enterspeed.Source.UmbracoCms.Base.Components
                 .To<AddContentStateToJobsTable>("enterspeedjobs-db-v3")
                 .To<AddFailedCountToJobsTable>("enterspeedjobs-db-v4");
 
-            var upgrader = new Upgrader(migrationPlan);
-            await upgrader.ExecuteAsync(_migrationPlanExecutor, _scopeProvider, _keyValueService);
+            return migrationPlan;
         }
-
-        public Task TerminateAsync(bool isRestarting, CancellationToken cancellationToken)
-        {
-            return Task.CompletedTask;
-        }
-#else
-        public void Initialize()
-        {
-            if (_runtimeState.Level < RuntimeLevel.Run)
-                return;
-
-            var migrationPlan = new MigrationPlan("EnterspeedJobs");
-            migrationPlan.From(string.Empty)
-                .To<EnterspeedJobsTableMigration>("enterspeedjobs-db")
-                .To<AddEntityTypeToJobsTable>("enterspeedjobs-db-v2")
-                .To<AddContentStateToJobsTable>("enterspeedjobs-db-v3")
-                .To<AddFailedCountToJobsTable>("enterspeedjobs-db-v4");
-
-            var upgrader = new Upgrader(migrationPlan);
-            upgrader.Execute(_migrationPlanExecutor, _scopeProvider, _keyValueService);
-        }
-
-        public void Terminate()
-        {
-        }
-#endif
     }
 }
