+++ src/Enterspeed.Source.UmbracoCms.Base/Data/Migration/AddFailedCountToJobsTable.cs
@@ -1,27 +1,16 @@
-#if NET10_0_OR_GREATER
-using System.Threading.Tasks;
-#endif
 using Enterspeed.Source.UmbracoCms.Base.Data.Schemas;
 using Umbraco.Cms.Infrastructure.Migrations;
 
 namespace Enterspeed.Source.UmbracoCms.Base.Data.Migration
 {
-#if NET10_0_OR_GREATER
-    public class AddFailedCountToJobsTable : AsyncMigrationBase
-#else
-    public class AddFailedCountToJobsTable : MigrationBase
-#endif
+    public class AddFailedCountToJobsTable : EnterspeedMigrationBase
     {
         public AddFailedCountToJobsTable(IMigrationContext context)
             : base(context)
         {
         }
 
-#if NET10_0_OR_GREATER
-        protected override Task MigrateAsync()
-#else
-        protected override void Migrate()
-#endif
+        protected override void ExecuteMigration()
         {
             var jobsTable = Database.PocoDataFactory.ForType(typeof(EnterspeedJobSchema));
             var jobsTableName = jobsTable.TableInfo.TableName;
@@ -37,10 +26,6 @@ namespace Enterspeed.Source.UmbracoCms.Base.Data.Migration
                     .AsInt32().NotNullable().WithDefaultValue(0)
                     .Do();
             }
-#if NET10_0_OR_GREATER
-
-            return Task.CompletedTask;
-#endif
         }
     }
 }
